# src/app/api/rsvp/route.ts
import { NextResponse, type NextRequest } from "next/server";
import { rsvpSchema } from "@/lib/validators";
import { supabaseAdmin } from "@/lib/supabase/admin";
import { sendRsvpEmail } from "@/lib/email";
import { rateLimit } from "@/lib/rate-limit";
import { LIMITS, RSVP_OPEN } from "@/lib/limits";
import { signQrToken } from "@/lib/qr-token";
import { ipFromHeaders } from "@/lib/audit";

export const runtime = "nodejs";

const IDEMPOTENCY_KEY_RE = /^[A-Za-z0-9_\-:.]{8,200}$/;

export async function POST(req: NextRequest) {
  if (!RSVP_OPEN) {
    return NextResponse.json(
      { error: "Inscrições encerradas." },
      { status: 503 },
    );
  }

  const ip = ipFromHeaders(req.headers) ?? "unknown";

  // ── Idempotency-Key (optional). Caller pode dar uma key estável (ex.: hash
  // local do form) e re-tentar com a mesma key sem disparar email duplicado.
  // Resposta cached é mantida 1h por (IP, key).
  const idempotencyKey = req.headers.get("idempotency-key");
  const supabase = supabaseAdmin();
  if (idempotencyKey) {
    if (!IDEMPOTENCY_KEY_RE.test(idempotencyKey)) {
      return NextResponse.json(
        { error: "Idempotency-Key inválida." },
        { status: 400 },
      );
    }
    const { data: cached } = await supabase
      .from("idempotency_keys")
      .select("response, status_code")
      .eq("scope", "rsvp")
      .eq("key", `${ip}:${idempotencyKey}`)
      .gt("expires_at", new Date().toISOString())
      .maybeSingle();
    if (cached) {
      return NextResponse.json(cached.response ?? {}, {
        status: cached.status_code ?? 200,
      });
    }
  }

  const body = await req.json().catch(() => null);
  const parsed = rsvpSchema.safeParse(body);

  if (!parsed.success) {
    return NextResponse.json(
      {
        error: "Dados inválidos.",
        issues: parsed.error.flatten().fieldErrors,
      },
      { status: 400 },
    );
  }

  const data = parsed.data;

  // Anti-spam tiered:
  //   1. per IP — blocks volumetric abuse from one source
  //   2. per (IP, email) — blocks retries of same submission
  //   3. per email globally — blocks IP-rotating abuse against same recipient
  //      (caps email vendor cost amplification)
  const ipRl = await rateLimit(
    `rsvp:ip:${ip}`,
    LIMITS.rsvp.perIp.max,
    LIMITS.rsvp.perIp.windowMs,
  );
  if (!ipRl.ok) {
    return NextResponse.json(
      { error: "Demasiados pedidos. Tenta mais tarde." },
      { status: 429, headers: { "Retry-After": String(ipRl.retryAfterSeconds) } },
    );
  }
  const rl = await rateLimit(
    `rsvp:${ip}:${data.email}`,
    LIMITS.rsvp.perIpEmail.max,
    LIMITS.rsvp.perIpEmail.windowMs,
  );
  if (!rl.ok) {
    return NextResponse.json(
      { error: "Demasiados pedidos. Tenta mais tarde." },
      { status: 429, headers: { "Retry-After": String(rl.retryAfterSeconds) } },
    );
  }
  const emailGlobalRl = await rateLimit(
    `rsvp:email:${data.email}`,
    LIMITS.rsvp.perEmailGlobal.max,
    LIMITS.rsvp.perEmailGlobal.windowMs,
  );
  if (!emailGlobalRl.ok) {
    // Generic message — does not confirm whether email already exists.
    return NextResponse.json(
      { error: "Demasiados pedidos. Tenta mais tarde." },
      { status: 429, headers: { "Retry-After": String(emailGlobalRl.retryAfterSeconds) } },
    );
  }

  const companion_count = data.acompanhante === "sim" ? 1 : 0;
  const companion_names =
    companion_count === 1 && data.companion_nome ? [data.companion_nome] : [];

  const { data: inserted, error: insertError } = await supabase
    .from("guests")
    .insert({
      name: data.name,
      email: data.email,
      phone: data.phone,
      companion_count,
      companion_names,
    })
    .select("id, token")
    .single();

  if (insertError) {
    if (insertError.code === "23505") {
      const dupeBody = { ok: true } as const;
      await cacheIdempotency(supabase, ip, idempotencyKey, dupeBody, 200);
      return NextResponse.json(dupeBody);
    }
    console.error("[rsvp] insert", insertError.code ?? "unknown");
    return NextResponse.json(
      { error: "Não foi possível gravar. Tenta novamente." },
      { status: 500 },
    );
  }

  // Token público assinado (ou UUID legacy se secret não definido).
  const publicToken = await signQrToken(inserted.token);

  try {
    await sendRsvpEmail({
      to: data.email,
      name: data.name,
      token: publicToken,
    });

    await supabase
      .from("guests")
      .update({ email_sent_at: new Date().toISOString() })
      .eq("id", inserted.id);
  } catch (e) {
    console.error("[rsvp] email", e);
    // Registo ficou gravado. Cron `/api/cron/email-retry` apanha o resto.
  }

  const responseBody = { token: publicToken };
  await cacheIdempotency(supabase, ip, idempotencyKey, responseBody, 200);
  return NextResponse.json(responseBody);
}

async function cacheIdempotency(
  supabase: ReturnType<typeof supabaseAdmin>,
  ip: string,
  key: string | null,
  response: Record<string, unknown>,
  statusCode: number,
) {
  if (!key) return;
  try {
    await supabase.from("idempotency_keys").insert({
      scope: "rsvp",
      key: `${ip}:${key}`,
      response,
      status_code: statusCode,
      expires_at: new Date(Date.now() + 60 * 60_000).toISOString(),
    });
  } catch {
    /* idempotency cache é best-effort */
  }
}